feat(metrics): add volume metrics, adherence endpoints, and nutrition schema fix
- New VolumeMetricsService with weekly volume and adherence rate calculations
- New /metrics/volume and /metrics/adherence endpoints
- Repository methods for volume aggregation and log counting by date range
- Changed nutrition plan target macros from int to float
- Added Makefile for dev convenience

+++ backend/app/schemas/nutrition_plan.py
@@ -8,10 +8,10 @@ from pydantic import BaseModel
 class NutritionPlanBase(BaseModel):
     name: str
     days: Optional[Any] = None  # JSON: list of day objects
-    target_calories: Optional[int] = None
-    target_protein: Optional[int] = None
-    target_carbs: Optional[int] = None
-    target_fat: Optional[int] = None
+    target_calories: Optional[float] = None
+    target_protein: Optional[float] = None
+    target_carbs: Optional[float] = None
+    target_fat: Optional[float] = None
     fiber_g: Optional[int] = None
     water_ml: Optional[int] = None
     meals_per_day: Optional[int] = None
